libserialize: Unconfigure tests during normal build
